refactor k8s api
