+++ src/integration/k8s_live/mod.rs
@@ -1,3 +1,4 @@
+pub mod k8s_api;
 pub mod k8s_integration_plugin;
 pub mod k8s_integration_systems;
 // pub mod k8s_utils;
